Inline implementations in actuals
With the dedicated platform testing gone, this indirection is no longer needed. The common tests still validate behavioral conformity across all platforms.
