+++ src/commonImplementation/kotlin/StringBuilderExtensions.kt
@@ -1,8 +1,15 @@
 package de.cketti.codepoints
 
+import de.cketti.codepoints.CodePoints.highSurrogate
+import de.cketti.codepoints.CodePoints.isBmpCodePoint
+import de.cketti.codepoints.CodePoints.lowSurrogate
 import kotlin.text.StringBuilder
-import de.cketti.codepoints.internal.appendCodePoint as commonAppendCodePoint
 
 actual fun StringBuilder.appendCodePoint(codePoint: Int): StringBuilder = apply {
-    commonAppendCodePoint(this, codePoint)
+    if (isBmpCodePoint(codePoint)) {
+        append(codePoint.toChar())
+    } else {
+        append(highSurrogate(codePoint))
+        append(lowSurrogate(codePoint))
+    }
 }
